import multi-page PDF: added markers to viewer but have more to learn there
autocount.html now allow a multiple page PDF to be dragged onto it.  It opens all pages and uses an abbreviated version of Potrace to find outlines, then sorts them by size, and adds counts for each symbol (within a similarity threshold) to a table and also overlays the symbols on the image allowing each to be highlighted in red as it is hovered over.

trying to integrate this into Mozilla's viewer was too difficult at this point, but I was able to identify some key areas where I could insert code into viewer.js